fix: check if depdnency graph file exists before removing

+++ packages/uikit-workshop/build-tools.js
@@ -3,7 +3,10 @@ const path = require('path');
 const fs = require('fs-extra');
 
 fs.mkdirp('./node_modules/@pattern-lab/');
-fs.unlinkSync('./dependencyGraph.json');
+// clear any previously generated dependency graph files if present
+if (fs.existsSync('./dependencyGraph.json')) {
+  fs.unlinkSync('./dependencyGraph.json');
+}
 fs.ensureSymlink(__dirname, './node_modules/@pattern-lab/uikit-workshop');
 
 const configKeysEndingWithASlash = [
